Add generic analytics check
Check if the project's website listed on GitHub is using any of the
analytics providers checked. At the moment the following are supported:

- Google Analytics 3 (Universal Analytics)
- Google Analytics 4
- HubSpot

In addition to displaying if the check passes or not, the providers
detected, which can be multiple, are also displayed in the UI, CSV
export and the console output of the linter CLI tool.

This check replaces the Google Analytics 4 check.

Closes #470

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>
Signed-off-by: Cintia Sanchez Garcia <cynthiasg@icloud.com>
Co-authored-by: Sergio Castaño Arteaga <tegioz@icloud.com>
Co-authored-by: Cintia Sanchez Garcia <cynthiasg@icloud.com>