Disable branch protection check temporarily (#528)
At the moment, this scorecard check requires a PAT token. This is not a
problem for the CLOMonitor tracker, which collects data for the
`clomonitor.io` site, as it uses PAT tokens.

However, when running the CLOMonitor linter from CI workflows, it is not
possible to use the automatic token authentication (GITHUB_TOKEN) and
some additional setup is required. This may have an impact on user
adoption, so we'll disable this check temporarily.

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>