Store and display check sets used to build report
When the repository's check sets field is updated, until it's tracked
again (max 1hr), we display the previous repository's report with the
new check sets. This may lead to confusions as for that short period
of time they won't match.

This change fixes that by storing and displaying the check sets used
when the report was created. As soon as the repository is tracked
again, the new check sets will be used to produce the report and they
will be displayed in the UI.

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>
Signed-off-by: Cintia Sanchez Garcia <cynthiasg@icloud.com>
Co-authored-by: Sergio Castaño Arteaga <tegioz@icloud.com>
Co-authored-by: Cintia Sanchez Garcia <cynthiasg@icloud.com>