Create snapshots of projects data periodically (#691)
A new component named `archiver` has been added to CLOMonitor. This
component takes snapshots of the projects data periodically and cleans
up snapshots that are no longer needed. Snapshots available are now
returned when fetching the project's data.

This is some ground work in preparation for the `Time Machine` feature.

Related to #681

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>