add more Metal3 repos to config (#1377)
Metal3 is preparing for the incubation and we want to update the key
repositories for the project in the config. Alongside CAPM3, BMO and
docs, we have another key repository IPAM as well as new community
repository, separated from the docs repository. We also have now
releases for ironic-image, and that can be checked at code-lite level.
BMO is full-fledged code repository and moves to code level checks.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>